(chore): temp change for mavonic demo

+++ core/glove_loader.py
@@ -3,10 +3,8 @@ import os
 import numpy as np
 from services.logger import get_logger
 
-
 logger = get_logger("glove")
 _glove_cache = None
-
 
 def load_glove(path: str = "data/glove.6B.300d.txt"):
     global _glove_cache
@@ -24,7 +22,7 @@ def load_glove(path: str = "data/glove.6B.300d.txt"):
             for line in f:
                 parts = line.strip().split()
                 word = parts[0]
-                vec = np.array(parts[1:], dtype=float)
+                vec = np.array(parts[1:], dtype=np.float32)  # use float32 for memory efficiency
                 embeddings[word] = vec
         joblib.dump(embeddings, cache_path)
         _glove_cache = embeddings
